make a reusable LifeDisplay component
